fix: add MainOpenCapsule component 수정
