--- frontend/src/components/MainOpenCapsule/styles.module.scss
@@ -1,7 +1,0 @@
-.capsule_box {
-  backgroundsize: cover;
-  objectfit: cover;
-  width: 5rem;
-  height: 6rem;
-  borderradius: 50% / 50%;
-}
